requirements.txt
langchain
pandasai
langchain-groq
langchain-experimental
streamlit
python-dotenv
pandasai